% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions_processing_tracked.R
\name{analyze_data}
\alias{analyze_data}
\title{Title}
\usage{
analyze_data(data, output_dir, save_as, radius, drop_contours, frame_rate = 16)
}
\arguments{
\item{data}{the data.frame which is the output of the function create_dataset_from_dir}

\item{output_dir}{the directory where the analysis should be saved}

\item{save_as}{the name under which the analysis csv-data should be saved}

\item{radius}{the radius of the petri dish}

\item{drop_contours}{if contours should be dropped (contours will be useful for the animation of larvae)}

\item{frame_rate}{the frame rate of the tracked videos}
}
\value{

}
\description{
Title
}
